add dashes to index page, refactor HomePageController
